requirements.txt
apache-libcloud
#easydict